requirements.txt
requests
pillow
gpiozero==1.6.5
pisugar
waveshare-epd @ git+https://github.com/waveshareteam/e-Paper.git#subdirectory=RaspberryPi_JetsonNano/python&egg=waveshare-epd
hitherdither @ git+https://www.github.com/hbldh/hitherdither